+++ crates/miden-protocol/src/lib.rs
@@ -60,23 +60,7 @@ pub mod crypto {
     pub use miden_crypto::{SequentialCommit, dsa, hash, ies, merkle, rand, utils};
 }
 
-pub mod utils {
-    pub use miden_core::utils::*;
-    pub use miden_crypto::utils::{HexParseError, bytes_to_hex_string, hex_to_bytes};
-    pub use miden_utils_sync as sync;
-
-    pub mod serde {
-        pub use miden_crypto::utils::{
-            BudgetedReader,
-            ByteReader,
-            ByteWriter,
-            Deserializable,
-            DeserializationError,
-            Serializable,
-            SliceReader,
-        };
-    }
-}
+pub mod utils;
 
 pub mod vm {
     pub use miden_assembly_syntax::ast::{AttributeSet, QualifiedProcedureName};